add ゲストログイン機能

+++ app/controllers/admin/users_controller.rb
@@ -1,4 +1,5 @@
 class Admin::UsersController < ApplicationController
+  before_action :authenticate_admin!
   def show
     @user = User.find(params[:id])
     @posts = @user.posts
@@ -9,8 +10,13 @@ class Admin::UsersController < ApplicationController
   end
 
   def destroy
-    user = User.find(params[:id])
-    user.destroy
-    redirect_to admin_root_path
+    @user = current_admin
+    if @user.email == 'guest@example.com'
+      redirect_to admin_root_path, notice: "ゲストログインユーザーでは削除できません"
+    else
+      user = User.find(params[:id])
+      user.destroy
+      redirect_to admin_root_path
+    end
   end
 end
